pyproject.toml
[project]
name = "lautoda"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "evotoolkit>=1.0.0",
    "torch>=2.0",
    "torchvision",
    "foolbox",
    "eagerpy",
    "robustbench @ git+https://github.com/RobustBench/robustbench.git",
    "numpy",
    "python-dotenv",
]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
